(refactor) rename cli folder
