+++ bin/dev_sample.py
@@ -21,14 +21,14 @@ from typing import (
 )
 
 from hummingbot import init_logging
-from hummingbot.cli.hummingbot_application import HummingbotApplication
-from hummingbot.cli.config.global_config_map import global_config_map
-from hummingbot.cli.config.in_memory_config_map import in_memory_config_map
-from hummingbot.cli.config.config_helpers import (
+from hummingbot.client.hummingbot_application import HummingbotApplication
+from hummingbot.client.config.global_config_map import global_config_map
+from hummingbot.client.config.in_memory_config_map import in_memory_config_map
+from hummingbot.client.config.config_helpers import (
     create_yml_files,
     read_configs_from_yml
 )
-from hummingbot.cli.ui.stdout_redirection import patch_stdout
+from hummingbot.client.ui.stdout_redirection import patch_stdout
 from hummingbot.core.utils.wallet_setup import unlock_wallet
