+++ hummingbot/client/config/global_config_map.py
@@ -1,7 +1,7 @@
 import random
 from typing import Callable
-from hummingbot.cli.config.config_var import ConfigVar
-from hummingbot.cli.settings import (
+from hummingbot.client.config.config_var import ConfigVar
+from hummingbot.client.settings import (
     required_exchanges,
     DEXES,
     DEFAULT_KEY_FILE_PATH,
