+++ hummingbot/client/errors.py
