+++ hummingbot/client/settings.py
