+++ hummingbot/client/ui/keybindings.py
