+++ hummingbot/client/ui/style.py
