+++ hummingbot/core/utils/stop_loss_tracker.py
@@ -8,7 +8,7 @@ from typing import (
 )
 
 from hummingbot.data_feed.data_feed_base import DataFeedBase
-from hummingbot.cli.config.global_config_map import global_config_map
+from hummingbot.client.config.global_config_map import global_config_map
 from hummingbot.market.market_base import MarketBase
 from hummingbot.logger import HummingbotLogger
