+++ hummingbot/strategy/cross_exchange_market_making/cross_exchange_market_making_config_map.py
@@ -1,9 +1,9 @@
-from hummingbot.cli.config.config_var import ConfigVar
-from hummingbot.cli.config.config_validators import (
+from hummingbot.client.config.config_var import ConfigVar
+from hummingbot.client.config.config_validators import (
     is_exchange,
     is_valid_market_symbol,
 )
-from hummingbot.cli.settings import (
+from hummingbot.client.settings import (
     required_exchanges,
     EXAMPLE_PAIRS,
 )
