+++ hummingbot/strategy/discovery/discovery_config_map.py
@@ -1,8 +1,8 @@
-from hummingbot.cli.config.config_var import ConfigVar
-from hummingbot.cli.config.config_validators import (
+from hummingbot.client.config.config_var import ConfigVar
+from hummingbot.client.config.config_validators import (
     is_exchange,
 )
-from hummingbot.cli.settings import (
+from hummingbot.client.settings import (
     EXAMPLE_PAIRS,
     required_exchanges,
 )
