+++ hummingbot/wallet/ethereum/erc20_tokens.json
