+++ setup.py
@@ -55,7 +55,7 @@ def main():
             "core/cpp/*",
             "wallet/ethereum/zero_ex/zero_ex_exchange_abi.json"
             "wallet/ethereum/token_abi/*.json"
-            "erc20_tokens.json",
+            "wallet/ethereum/erc20_tokens.json"
             "VERSION",
             "templates/*TEMPLATE.yml"
         ],
